Renamed admins route to admin
Corresponding views/controllers have been refactored
accordingly.
