+++ app/controllers/admin/dashboard_controller.rb
@@ -1,0 +1,4 @@
+class Admin::DashboardController < Admin::BaseController
+  def show
+  end
+end
