+++ app/controllers/admin/invitations_controller.rb
@@ -1,0 +1,3 @@
+class Admin::InvitationsController < Devise::InvitationsController
+  before_action :authorize_admin!
+end
