--- app/controllers/admins/base_controller.rb
@@ -1,12 +1,0 @@
-module Admins
-  class BaseController < ApplicationController
-    before_action :authenticate_admin!
-    before_action :authorize_admin!
-
-    private
-
-    def authorize_admin!
-      authorize! :manage, :admin_dashboard
-    end
-  end
-end
