--- app/controllers/admins/dashboard_controller.rb
@@ -1,6 +1,0 @@
-module Admins
-  class DashboardController < BaseController
-    def show
-    end
-  end
-end
