--- app/controllers/admins/invitations_controller.rb
@@ -1,5 +1,0 @@
-module Admins
-  class InvitationsController < Devise::InvitationsController
-    before_action :authorize_admin!
-  end
-end
