+++ app/views/admin/dashboard/show.html.slim
