+++ app/views/admin/semesters/show.html.slim
