+++ app/views/admin/sessions/new.html.slim
@@ -12,7 +12,7 @@
         = f.input :email, required: false, autofocus: true
         = f.input :password, required: false
         = f.input :remember_me, as: :boolean if devise_mapping.rememberable?
-      = render "admins/shared/links"
+      = render "admin/shared/links"
       .spacer-small
       .form-actions
         = f.button :submit, "Log in"
