+++ app/views/admin/shared/_links.html.slim
