+++ app/views/admin/student_applications/index.html.slim
