+++ app/views/admin/student_applications/show.html.slim
